upgrade-spring-boot
Upgrade spring boot to 3.3.7
Add PostController

+++ src/main/java/com/brain/history/multimedia/model/PostView.java
@@ -12,11 +12,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PostView {
-
     private Long collectionId;
     private String postDate;
     private String overview;
-    private Long array;
+    private Integer array;
     private List<PostedContent> content = new ArrayList<>();
 
     public List<PostedContent> add(PostedContent postedContent) {
